# api-docs.yaml
openapi: 3.0.3
info:
  title: Product Cache API
  description: >
    A hands-on API for learning and comparing the four main cache patterns:
    Cache-Aside, Read-Through, Write-Through, and Write-Behind.
    Each endpoint is scoped to a single pattern — no endpoint mixes two cache strategies
    unless explicitly noted and justified.
  version: 1.0.0
  contact:
    name: Product Cache API

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Cache-Aside
    description: >
      The application manually controls the cache.
      It checks Redis first; on a miss, it queries the DB and populates Redis itself.
      The DB has no awareness of the cache.
  - name: Read-Through
    description: >
      The cache sits in front of the DB. The application always talks to the cache.
      On a miss, Spring's @Cacheable proxy loads from the DB and stores the result automatically.
      The caller never writes to Redis directly.
  - name: Write-Through
    description: >
      Every write is applied to the DB and Redis synchronously in the same operation.
      The cache is always consistent after a write. Trade-off: higher write latency.
  - name: Write-Back/Write-Behind
    description: >
      The application writes to Redis immediately and returns.
      The DB write is deferred to a background async process.
      Lowest write latency. Trade-off: potential data loss if the app crashes before the flush.


paths:

  # -----------------------------------------------------------------------
  # CACHE-ASIDE
  # -----------------------------------------------------------------------

  /products/{id}:
    get:
      tags:
        - Cache-Aside
      summary: Get product by ID
      description: >
        **Pattern: Cache-Aside**

        The application checks Redis for the key `product:{id}`.

        - **HIT**: deserializes and returns the cached value. The DB is never touched.

        - **MISS**: queries H2, writes the result to Redis with a 3-minute TTL,
          then returns.

        The Redis read/write logic is written explicitly in the service layer —
        no Spring abstraction is used here. This is the key contrast with Read-Through.
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Write-Through
      summary: Update product
      description: >
        **Pattern: Write-Through**

        Updates the product in H2 and immediately overwrites `product:{id}` in Redis
        within the same synchronous operation. The TTL is reset to 3 minutes.

        After this call returns, any subsequent GET will find the fresh value in Redis.
        Stale reads are not possible with this pattern.
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated. DB and cache are now in sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /products/:
    get:
      tags:
        - Cache-Aside
      summary: List all products
      description: >
        **Pattern: Cache-Aside — applied to a collection**

        Retrieves all products from the database and returns them.
        This endpoint demonstrates how Cache-Aside can be applied to collection queries.
      operationId: listProducts
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'

    post:
      tags:
        - Write-Through
      summary: Create product
      description: >
        **Pattern: Write-Through**

        Inserts the product into H2 and simultaneously writes the result to Redis
        under `product:{id}` with a 3-minute TTL, in the same synchronous operation.

        This is the creation counterpart to PUT /products/{id}: together they cover the full
        Write-Through lifecycle (insert and update).
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created. Written to DB and Redis.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # -----------------------------------------------------------------------
  # READ-THROUGH
  # -----------------------------------------------------------------------

  /products/{id}/read-through:
    get:
      tags:
        - Read-Through
      summary: Get product via @Cacheable (Read-Through)
      description: >
        **Pattern: Read-Through**

        Functionally equivalent to GET /products/{id}, but implemented using
        Spring's `@Cacheable("products")` annotation instead of manual Redis calls.

        On a cache miss, Spring's AOP proxy calls the method body (the DB query)
        and stores the result in Redis automatically. The service method contains
        zero Redis code — it is pure business logic.

        This endpoint exists separately from GET /products/{id} so you can compare
        the two implementations side-by-side:
        - Cache-Aside: you write `redis.get`, `redis.setEx` explicitly
        - Read-Through: Spring's proxy does it transparently via `@Cacheable`

        Both use Redis as the cache store. The pattern difference is in who manages
        the cache population logic.
      operationId: getProductReadThrough
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product found (from cache or DB, transparently)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # -----------------------------------------------------------------------
  # WRITE-BACK/WRITE-BEHIND
  # -----------------------------------------------------------------------

  /products/{id}/write-back:
    put:
      tags:
        - Write-Back/Write-Behind
      summary: Update product asynchronously (Write-Back/Write-Behind)
      description: >
        **Pattern: Write-Back/Write-Behind**

        Updates Redis immediately and returns `202 Accepted` before the DB is written.
        The DB update is dispatched to a background thread via `@Async` and executes
        approximately 2 seconds after the response is returned.

        The 202 status code is intentional: it signals that the request was accepted
        but persistence is not yet complete. This is the observable contract of Write-Back/Write-Behind.

        This pattern provides the lowest write latency since the application doesn't wait
        for the DB write to complete before responding to the client.
      operationId: updateProductWriteBack
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '202':
          description: >
            Accepted. Redis updated immediately. DB write is pending in background.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /products/{id}/write-back-queue:
    post:
      tags:
        - Write-Back/Write-Behind
      summary: Update product via queue (Write-Back/Write-Behind with Queue)
      description: >
        **Pattern: Write-Back/Write-Behind with Queue**

        Updates Redis immediately and enqueues the DB write operation for background processing.
        Returns `201 Created` to indicate the resource was successfully created/updated in cache.

        This variant uses a queue-based approach for the async DB write, providing better
        reliability and ordering guarantees compared to simple async execution.
        The queue ensures that writes are processed in order and can be retried on failure.

        The 201 status code indicates that the resource was successfully created/updated
        in the cache. The DB persistence happens asynchronously in the background.
      operationId: updateProductWriteBackQueue
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: >
            Created. Redis updated immediately. DB write is queued for background processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

# -----------------------------------------------------------------------
# COMPONENTS
# -----------------------------------------------------------------------

components:

  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: Unique product identifier
      schema:
        type: integer
        format: int64
        example: 1

  schemas:

    ProductRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Mechanical Keyboard
        price:
          type: number
          format: decimal
          minimum: 0.01
          example: 249.99
        category:
          type: string
          maxLength: 50
          example: Electronics
        stock:
          type: integer
          minimum: 0
          example: 42

    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Mechanical Keyboard
        price:
          type: number
          format: decimal
          example: 249.99
        category:
          type: string
          example: Electronics
        stock:
          type: integer
          example: 42
        createdAt:
          type: string
          format: date-time
          example: "2025-03-12T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-03-12T11:30:00Z"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Product with id 99 not found
        timestamp:
          type: string
          format: date-time
          example: "2025-03-12T11:30:00Z"

  responses:
    NotFound:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'